requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
openai==1.63.2
python-dotenv
regex
pyyaml
flask
